refactor(diff): set diff base in new window rather than origin
Changes the creation order of main, diff base, info buffers.
The main tardis buffer, showing the current revision, will always be in the origin window, regardless of whether the diff base is visible in a split window.
This is to make managing the diff base window more straightforward.
The diff and info buffers, if enabled, are created after `Session:goto_buffer` is called on `Session:init`. This requires moving the initial `goto_buffer` call from the end of `SessionManager:create_session` into `Session:init`